test: rename the test and use decimal

+++ Shops/Services/IShopManager.cs
@@ -5,7 +5,7 @@ namespace Shops.Services
 {
     public interface IShopManager
     {
-        Customer AddCustomer(string name, int money);
+        Customer AddCustomer(string name, decimal money);
         Shop AddShop(string name, Adress adress);
         Product GetProduct(Shop shop, Product product);
 
@@ -14,7 +14,7 @@ namespace Shops.Services
         Customer AddCustomersProduct(Customer customer, Product product);
         Shop FindShop(int id);
         Customer FindCustomer(string name);
-        Product ChangePriceProduct(Shop shop, Product product, int newPrice);
+        Product ChangePriceProduct(Shop shop, Product product, decimal newPrice);
         Shop FindShopWithCheapestProducts(List<Product> products);
 
         void PurchaseGoods(Shop shop, Customer customer);
